requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
databricks-sdk==0.58.0
dataclasses-json==0.6.7
Deprecated==1.2.18
docker==7.1.0
easyocr==1.7.1
fastapi==0.116.0
filelock==3.16.1
Flask==3.1.0
flask-cors==6.0.0
Flask-Limiter==3.12
flatbuffers==24.3.25
fonttools==4.58.0
frozenlist==1.7.0
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
grpcio==1.68.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.6.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
lazy_loader==0.4
libclang==18.1.1
limits==5.4.0
lxml==5.3.2
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.4.1
mlflow==3.1.1
mlflow-skinny==3.1.1
mpmath==1.3.0
msal==1.32.3
multidict==6.6.3
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.1
numpy==2.0.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
propcache==0.3.2
protobuf==5.28.3
psutil==7.0.0
psycopg2==2.9.10
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.3
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
statsmodels==0.14.4
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.67.1
transformers==4.53.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
ultralytics==8.1.9
urllib3==2.2.3
uvicorn==0.35.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.0
xgboost==2.1.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
psycopg2-binary

